==> prednisolon/02_track.Rmd <==
---
title: "Prepae metadata file for asthma data"
author: "TB"
date: "7/28/2017"
output: html_document
---
```{r}
library("tidyverse")
library("stringr")
library("readr")
library("plotly")
library("ggplot2")
library("neutrominer")
source("../utilities.r")
```

# Load data
The analysis folder contains the segmentation and tracking results. Each 
time lapse experiment is stored in separate folder.
```{r}
analysis_folder <- "../../../analysis/20171006/"
csv_single_cells <- "Tracked_detectedNeutros.csv"
```

# Output 
Results of neutrominer::track are stored in csv_tracks
```{r}
csv_tracks <- 'tracks_cp.csv'
```


# Metadata 
```{r}
strata <- c('Metadata_condition','Metadata_date','Metadata_dose','Metadata_id')
```


# Parameter
To recalculate the speed of the neutrophils in µm/minute we need the pixel size
and the number of frames per minute.
```{r}
# scale speed from pixel/frame to µm/minute
pixel_size = 0.7819
frames_per_second = 2/60
# frames are recorded each 30 seconds, i.e. 
speed_scale <- pixel_size * frames_per_second

track_label = c('TrackObjects_Label')
```


# Load data 
The output of CellProfiler track module is loaded as CSV, tracked using cytominer::track and 
stored in the data folder as "track.csv". 
```{r}
# load csv files from all sub folders 
df <- read_csv(file.path(analysis_folder, csv_single_cells))

#df %>% 
#  mutate(Metadata_timePoint = Metadata_frame) %>%
#  dplyr::mutate(Metadata_timePoint = as.double(Metadata_timePoint)) 
```
```{r}
df %>% 
  ungroup() %>% 
  group_by(Metadata_condition) %>% 
  summarise(n = n())
```


# track the neutros
```{r}
tracks <- df %>%
  group_by_(.dots =  c(strata,track_label)) %>%
  rename(Metadata_timePoint = Metadata_frame) %>%
  arrange(Metadata_timePoint) %>%
  filter(Metadata_timePoint >19) %>%
  neutrominer::track(., c(strata,track_label))
```


## PostProcessing 
Calculate speed in µm/minute
```{r}
postprocessed <- tracks %>%
  mutate(Track_Speed = Track_Speed * speed_scale) %>%
  mutate(Track_Speed_max = Track_Speed_max * speed_scale) %>%
  mutate(Track_Speed_X = Track_Speed_X * speed_scale) %>%
  mutate(Track_Speed_Y = Track_Speed_Y * speed_scale) %>%
  mutate(Track_Distance_Traveled = pixel_size * Track_Distance_Traveled) %>%
  mutate(Track_Integrated_Distance_Traveled = pixel_size * Track_Integrated_Distance_Traveled) 
```

# Add Metadata_matrix
```{r}
# add Metadata_matrix
df_export <- add_metadata_matrix(postprocessed)
strata <- c("Metadata_matrix", strata)
```

# Save data 
* Correct Metadata: all na values are set to 'None'
* store data 
```{r}
df_export$Metadata_dose[is.na(df_export$Metadata_dose)] <- 'None'
write_csv(df_export, file.path(analysis_folder, csv_tracks))
```
# analyse quality
```{r}
 quality <- df_export %>% 
  group_by_(.dots = strata) %>%
    neutrominer::assess(.,min_path_length = 19) %>%
  arrange(VOT) %>%
  print
```
# 

```{r}
quality %>% 
  ungroup() %>% 
  filter(Metadata_condition %in% c("vehi","pred")) %>%
  group_by(Metadata_condition,Metadata_matrix) %>%
  summarise(mean_vot = mean(VOT), sd_vot = sd(VOT), n = n()) %>%
  print
```



```{r}
ggplot(data = quality) +
  aes(x = VOT, y = Exp_Mean_Track_Length, color = Metadata_condition) +
  geom_point()
```
